+++ geoip2/_internal.py
@@ -1,22 +1,21 @@
-"""This package contains internal utilities"""
+"""This package contains internal utilities."""
 
 # pylint: disable=too-few-public-methods
 from abc import ABCMeta
-from typing import Any
 
 
 class Model(metaclass=ABCMeta):
-    """Shared methods for MaxMind model classes"""
+    """Shared methods for MaxMind model classes."""
 
-    def __eq__(self, other: Any) -> bool:
+    def __eq__(self, other: object) -> bool:
         return isinstance(other, self.__class__) and self.to_dict() == other.to_dict()
 
-    def __ne__(self, other):
+    def __ne__(self, other) -> bool:
         return not self.__eq__(other)
 
     # pylint: disable=too-many-branches
-    def to_dict(self):
-        """Returns a dict of the object suitable for serialization"""
+    def to_dict(self) -> dict:
+        """Returns a dict of the object suitable for serialization."""
         result = {}
         for key, value in self.__dict__.items():
             if key.startswith("_"):