+++ geoip2/database.py
@@ -7,41 +7,41 @@ GeoIP2 Database Reader
 
 import inspect
 import os
-from typing import Any, AnyStr, cast, IO, Optional, Sequence, Type, Union
+from collections.abc import Sequence
+from typing import IO, Any, AnyStr, Optional, Union, cast
 
 import maxminddb
-
 from maxminddb import (
     MODE_AUTO,
-    MODE_MMAP,
-    MODE_MMAP_EXT,
+    MODE_FD,
     MODE_FILE,
     MODE_MEMORY,
-    MODE_FD,
+    MODE_MMAP,
+    MODE_MMAP_EXT,
 )
 
 import geoip2
-import geoip2.models
 import geoip2.errors
-from geoip2.types import IPAddress
+import geoip2.models
 from geoip2.models import (
     ASN,
+    ISP,
     AnonymousIP,
     City,
     ConnectionType,
     Country,
     Domain,
     Enterprise,
-    ISP,
 )
+from geoip2.types import IPAddress
 
 __all__ = [
     "MODE_AUTO",
-    "MODE_MMAP",
-    "MODE_MMAP_EXT",
+    "MODE_FD",
     "MODE_FILE",
     "MODE_MEMORY",
-    "MODE_FD",
+    "MODE_MMAP",
+    "MODE_MMAP_EXT",
     "Reader",
 ]
 
@@ -135,9 +135,9 @@ class Reader:
         :returns: :py:class:`geoip2.models.Country` object
 
         """
-
         return cast(
-            Country, self._model_for(geoip2.models.Country, "Country", ip_address)
+            Country,
+            self._model_for(geoip2.models.Country, "Country", ip_address),
         )
 
     def city(self, ip_address: IPAddress) -> City:
@@ -161,7 +161,9 @@ class Reader:
         return cast(
             AnonymousIP,
             self._flat_model_for(
-                geoip2.models.AnonymousIP, "GeoIP2-Anonymous-IP", ip_address
+                geoip2.models.AnonymousIP,
+                "GeoIP2-Anonymous-IP",
+                ip_address,
             ),
         )
 
@@ -174,7 +176,8 @@ class Reader:
 
         """
         return cast(
-            ASN, self._flat_model_for(geoip2.models.ASN, "GeoLite2-ASN", ip_address)
+            ASN,
+            self._flat_model_for(geoip2.models.ASN, "GeoLite2-ASN", ip_address),
         )
 
     def connection_type(self, ip_address: IPAddress) -> ConnectionType:
@@ -188,7 +191,9 @@ class Reader:
         return cast(
             ConnectionType,
             self._flat_model_for(
-                geoip2.models.ConnectionType, "GeoIP2-Connection-Type", ip_address
+                geoip2.models.ConnectionType,
+                "GeoIP2-Connection-Type",
+                ip_address,
             ),
         )
 
@@ -227,7 +232,8 @@ class Reader:
 
         """
         return cast(
-            ISP, self._flat_model_for(geoip2.models.ISP, "GeoIP2-ISP", ip_address)
+            ISP,
+            self._flat_model_for(geoip2.models.ISP, "GeoIP2-ISP", ip_address),
         )
 
     def _get(self, database_type: str, ip_address: IPAddress) -> Any:
@@ -247,19 +253,26 @@ class Reader:
 
     def _model_for(
         self,
-        model_class: Union[Type[Country], Type[Enterprise], Type[City]],
+        model_class: Union[type[Country], type[Enterprise], type[City]],
         types: str,
         ip_address: IPAddress,
     ) -> Union[Country, Enterprise, City]:
         (record, prefix_len) = self._get(types, ip_address)
         return model_class(
-            self._locales, ip_address=ip_address, prefix_len=prefix_len, **record
+            self._locales,
+            ip_address=ip_address,
+            prefix_len=prefix_len,
+            **record,
         )
 
     def _flat_model_for(
         self,
         model_class: Union[
-            Type[Domain], Type[ISP], Type[ConnectionType], Type[ASN], Type[AnonymousIP]
+            type[Domain],
+            type[ISP],
+            type[ConnectionType],
+            type[ASN],
+            type[AnonymousIP],
         ],
         types: str,
         ip_address: IPAddress,
@@ -278,5 +291,4 @@ class Reader:
 
     def close(self) -> None:
         """Closes the GeoIP2 database."""
-
         self._db_reader.close()