+++ tests/webservice_test.py
@@ -1,17 +1,17 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 import asyncio
 import copy
 import ipaddress
 import sys
-from typing import cast, Dict
 import unittest
-from pytest_httpserver import HeaderValueMatcher
-import pytest_httpserver
-import pytest
+from abc import ABC, abstractmethod
 from collections import defaultdict
+from typing import cast, Callable, Union
 
+import pytest
+import pytest_httpserver
+from pytest_httpserver import HeaderValueMatcher
 
 sys.path.append("..")
 import geoip2
@@ -27,7 +27,10 @@ from geoip2.errors import (
 from geoip2.webservice import AsyncClient, Client
 
 
-class TestBaseClient(unittest.TestCase):
+class TestBaseClient(unittest.TestCase, ABC):
+    client: Union[AsyncClient, Client]
+    client_class: Callable[[int, str], Union[AsyncClient, Client]]
+
     country = {
         "continent": {"code": "NA", "geoname_id": 42, "names": {"en": "North America"}},
         "country": {
@@ -51,9 +54,12 @@ class TestBaseClient(unittest.TestCase):
 
     # this is not a comprehensive representation of the
     # JSON from the server
-    insights = cast(Dict, copy.deepcopy(country))
+    insights = cast(dict, copy.deepcopy(country))
     insights["traits"]["user_count"] = 2
     insights["traits"]["static_ip_score"] = 1.3
+
+    @abstractmethod
+    def run_client(self, v): ...
 
     def _content_type(self, endpoint):
         return (
@@ -63,12 +69,13 @@ class TestBaseClient(unittest.TestCase):
         )
 
     @pytest.fixture(autouse=True)
-    def setup_httpserver(self, httpserver: pytest_httpserver.HTTPServer):
+    def setup_httpserver(self, httpserver: pytest_httpserver.HTTPServer) -> None:
         self.httpserver = httpserver
 
-    def test_country_ok(self):
+    def test_country_ok(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.2.3.4", method="GET"
+            "/geoip/v2.1/country/1.2.3.4",
+            method="GET",
         ).respond_with_json(
             self.country,
             status=200,
@@ -76,12 +83,16 @@ class TestBaseClient(unittest.TestCase):
         )
         country = self.run_client(self.client.country("1.2.3.4"))
         self.assertEqual(
-            type(country), geoip2.models.Country, "return value of client.country"
+            type(country),
+            geoip2.models.Country,
+            "return value of client.country",
         )
         self.assertEqual(country.continent.geoname_id, 42, "continent geoname_id is 42")
         self.assertEqual(country.continent.code, "NA", "continent code is NA")
         self.assertEqual(
-            country.continent.name, "North America", "continent name is North America"
+            country.continent.name,
+            "North America",
+            "continent name is North America",
         )
         self.assertEqual(country.country.geoname_id, 1, "country geoname_id is 1")
         self.assertIs(
@@ -91,7 +102,9 @@ class TestBaseClient(unittest.TestCase):
         )
         self.assertEqual(country.country.iso_code, "US", "country iso_code is US")
         self.assertEqual(
-            country.country.names, {"en": "United States of America"}, "country names"
+            country.country.names,
+            {"en": "United States of America"},
+            "country names",
         )
         self.assertEqual(
             country.country.name,
@@ -99,7 +112,9 @@ class TestBaseClient(unittest.TestCase):
             "country name is United States of America",
         )
         self.assertEqual(
-            country.maxmind.queries_remaining, 11, "queries_remaining is 11"
+            country.maxmind.queries_remaining,
+            11,
+            "queries_remaining is 11",
         )
         self.assertIs(
             country.registered_country.is_in_european_union,
@@ -107,14 +122,17 @@ class TestBaseClient(unittest.TestCase):
             "registered_country is_in_european_union is True",
         )
         self.assertEqual(
-            country.traits.network, ipaddress.ip_network("1.2.3.0/24"), "network"
+            country.traits.network,
+            ipaddress.ip_network("1.2.3.0/24"),
+            "network",
         )
         self.assertTrue(country.traits.is_anycast)
         self.assertEqual(country.to_dict(), self.country, "raw response is correct")
 
-    def test_me(self):
+    def test_me(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/me", method="GET"
+            "/geoip/v2.1/country/me",
+            method="GET",
         ).respond_with_json(
             self.country,
             status=200,
@@ -122,7 +140,9 @@ class TestBaseClient(unittest.TestCase):
         )
         implicit_me = self.run_client(self.client.country())
         self.assertEqual(
-            type(implicit_me), geoip2.models.Country, "country() returns Country object"
+            type(implicit_me),
+            geoip2.models.Country,
+            "country() returns Country object",
         )
         explicit_me = self.run_client(self.client.country())
         self.assertEqual(
@@ -131,9 +151,10 @@ class TestBaseClient(unittest.TestCase):
             "country('me') returns Country object",
         )
 
-    def test_200_error(self):
+    def test_200_error(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.1.1.1", method="GET"
+            "/geoip/v2.1/country/1.1.1.1",
+            method="GET",
         ).respond_with_data(
             "",
             status=200,
@@ -141,32 +162,37 @@ class TestBaseClient(unittest.TestCase):
         )
 
         with self.assertRaisesRegex(
-            GeoIP2Error, "could not decode the response as JSON"
+            GeoIP2Error,
+            "could not decode the response as JSON",
         ):
             self.run_client(self.client.country("1.1.1.1"))
 
-    def test_bad_ip_address(self):
+    def test_bad_ip_address(self) -> None:
         with self.assertRaisesRegex(
-            ValueError, "'1.2.3' does not appear to be an IPv4 or IPv6 address"
+            ValueError,
+            "'1.2.3' does not appear to be an IPv4 or IPv6 address",
         ):
             self.run_client(self.client.country("1.2.3"))
 
-    def test_no_body_error(self):
+    def test_no_body_error(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.2.3.7", method="GET"
+            "/geoip/v2.1/country/1.2.3.7",
+            method="GET",
         ).respond_with_data(
             "",
             status=400,
             content_type=self._content_type("country"),
         )
         with self.assertRaisesRegex(
-            HTTPError, "Received a 400 error for .* with no body"
+            HTTPError,
+            "Received a 400 error for .* with no body",
         ):
             self.run_client(self.client.country("1.2.3.7"))
 
-    def test_weird_body_error(self):
+    def test_weird_body_error(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.2.3.8", method="GET"
+            "/geoip/v2.1/country/1.2.3.8",
+            method="GET",
         ).respond_with_json(
             {"wierd": 42},
             status=400,
@@ -179,22 +205,25 @@ class TestBaseClient(unittest.TestCase):
         ):
             self.run_client(self.client.country("1.2.3.8"))
 
-    def test_bad_body_error(self):
+    def test_bad_body_error(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.2.3.9", method="GET"
+            "/geoip/v2.1/country/1.2.3.9",
+            method="GET",
         ).respond_with_data(
             "bad body",
             status=400,
             content_type=self._content_type("country"),
         )
         with self.assertRaisesRegex(
-            HTTPError, "it did not include the expected JSON body"
+            HTTPError,
+            "it did not include the expected JSON body",
         ):
             self.run_client(self.client.country("1.2.3.9"))
 
-    def test_500_error(self):
+    def test_500_error(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.2.3.10", method="GET"
+            "/geoip/v2.1/country/1.2.3.10",
+            method="GET",
         ).respond_with_data(
             "",
             status=500,
@@ -203,80 +232,84 @@ class TestBaseClient(unittest.TestCase):
         with self.assertRaisesRegex(HTTPError, r"Received a server error \(500\) for"):
             self.run_client(self.client.country("1.2.3.10"))
 
-    def test_300_error(self):
+    def test_300_error(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.2.3.11", method="GET"
+            "/geoip/v2.1/country/1.2.3.11",
+            method="GET",
         ).respond_with_data(
             "",
             status=300,
             content_type=self._content_type("country"),
         )
         with self.assertRaisesRegex(
-            HTTPError, r"Received a very surprising HTTP status \(300\) for"
+            HTTPError,
+            r"Received a very surprising HTTP status \(300\) for",
         ):
             self.run_client(self.client.country("1.2.3.11"))
 
-    def test_ip_address_required(self):
+    def test_ip_address_required(self) -> None:
         self._test_error(400, "IP_ADDRESS_REQUIRED", InvalidRequestError)
 
-    def test_ip_address_not_found(self):
+    def test_ip_address_not_found(self) -> None:
         self._test_error(404, "IP_ADDRESS_NOT_FOUND", AddressNotFoundError)
 
-    def test_ip_address_reserved(self):
+    def test_ip_address_reserved(self) -> None:
         self._test_error(400, "IP_ADDRESS_RESERVED", AddressNotFoundError)
 
-    def test_permission_required(self):
+    def test_permission_required(self) -> None:
         self._test_error(403, "PERMISSION_REQUIRED", PermissionRequiredError)
 
-    def test_auth_invalid(self):
+    def test_auth_invalid(self) -> None:
         self._test_error(400, "AUTHORIZATION_INVALID", AuthenticationError)
 
-    def test_license_key_required(self):
+    def test_license_key_required(self) -> None:
         self._test_error(401, "LICENSE_KEY_REQUIRED", AuthenticationError)
 
-    def test_account_id_required(self):
+    def test_account_id_required(self) -> None:
         self._test_error(401, "ACCOUNT_ID_REQUIRED", AuthenticationError)
 
-    def test_user_id_required(self):
+    def test_user_id_required(self) -> None:
         self._test_error(401, "USER_ID_REQUIRED", AuthenticationError)
 
-    def test_account_id_unkown(self):
+    def test_account_id_unkown(self) -> None:
         self._test_error(401, "ACCOUNT_ID_UNKNOWN", AuthenticationError)
 
-    def test_user_id_unkown(self):
+    def test_user_id_unkown(self) -> None:
         self._test_error(401, "USER_ID_UNKNOWN", AuthenticationError)
 
-    def test_out_of_queries_error(self):
+    def test_out_of_queries_error(self) -> None:
         self._test_error(402, "OUT_OF_QUERIES", OutOfQueriesError)
 
-    def _test_error(self, status, error_code, error_class):
+    def _test_error(self, status, error_code, error_class) -> None:
         msg = "Some error message"
         body = {"error": msg, "code": error_code}
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/1.2.3.18", method="GET"
+            "/geoip/v2.1/country/1.2.3.18",
+            method="GET",
         ).respond_with_json(
             body,
             status=status,
             content_type=self._content_type("country"),
         )
-        with self.assertRaisesRegex(error_class, msg):
+        with pytest.raises(error_class, match=msg):
             self.run_client(self.client.country("1.2.3.18"))
 
-    def test_unknown_error(self):
+    def test_unknown_error(self) -> None:
         msg = "Unknown error type"
         ip = "1.2.3.19"
         body = {"error": msg, "code": "UNKNOWN_TYPE"}
         self.httpserver.expect_request(
-            "/geoip/v2.1/country/" + ip, method="GET"
+            "/geoip/v2.1/country/" + ip,
+            method="GET",
         ).respond_with_json(
             body,
             status=400,
             content_type=self._content_type("country"),
         )
-        with self.assertRaisesRegex(InvalidRequestError, msg):
+        with pytest.raises(InvalidRequestError, match=msg):
             self.run_client(self.client.country(ip))
 
-    def test_request(self):
+    def test_request(self) -> None:
         def user_agent_compare(actual: str, expected: str) -> bool:
             if actual is None:
                 return False
@@ -293,8 +326,8 @@ class TestBaseClient(unittest.TestCase):
             header_value_matcher=HeaderValueMatcher(
                 defaultdict(
                     lambda: HeaderValueMatcher.default_header_value_matcher,
-                    {"User-Agent": user_agent_compare},
-                )
+                    {"User-Agent": user_agent_compare},  # type: ignore[dict-item]
+                ),
             ),
         ).respond_with_json(
             self.country,
@@ -303,9 +336,10 @@ class TestBaseClient(unittest.TestCase):
         )
         self.run_client(self.client.country("1.2.3.4"))
 
-    def test_city_ok(self):
+    def test_city_ok(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/city/1.2.3.4", method="GET"
+            "/geoip/v2.1/city/1.2.3.4",
+            method="GET",
         ).respond_with_json(
             self.country,
             status=200,
@@ -314,13 +348,16 @@ class TestBaseClient(unittest.TestCase):
         city = self.run_client(self.client.city("1.2.3.4"))
         self.assertEqual(type(city), geoip2.models.City, "return value of client.city")
         self.assertEqual(
-            city.traits.network, ipaddress.ip_network("1.2.3.0/24"), "network"
+            city.traits.network,
+            ipaddress.ip_network("1.2.3.0/24"),
+            "network",
         )
         self.assertTrue(city.traits.is_anycast)
 
-    def test_insights_ok(self):
+    def test_insights_ok(self) -> None:
         self.httpserver.expect_request(
-            "/geoip/v2.1/insights/1.2.3.4", method="GET"
+            "/geoip/v2.1/insights/1.2.3.4",
+            method="GET",
         ).respond_with_json(
             self.insights,
             status=200,
@@ -328,32 +365,39 @@ class TestBaseClient(unittest.TestCase):
         )
         insights = self.run_client(self.client.insights("1.2.3.4"))
         self.assertEqual(
-            type(insights), geoip2.models.Insights, "return value of client.insights"
+            type(insights),
+            geoip2.models.Insights,
+            "return value of client.insights",
         )
         self.assertEqual(
-            insights.traits.network, ipaddress.ip_network("1.2.3.0/24"), "network"
+            insights.traits.network,
+            ipaddress.ip_network("1.2.3.0/24"),
+            "network",
         )
         self.assertTrue(insights.traits.is_anycast)
         self.assertEqual(insights.traits.static_ip_score, 1.3, "static_ip_score is 1.3")
         self.assertEqual(insights.traits.user_count, 2, "user_count is 2")
 
-    def test_named_constructor_args(self):
+    def test_named_constructor_args(self) -> None:
         id = 47
         key = "1234567890ab"
-        client = self.client_class(account_id=id, license_key=key)
+        client = self.client_class(id, key)
         self.assertEqual(client._account_id, str(id))
         self.assertEqual(client._license_key, key)
 
-    def test_missing_constructor_args(self):
+    def test_missing_constructor_args(self) -> None:
         with self.assertRaises(TypeError):
-            self.client_class(license_key="1234567890ab")
+
+            self.client_class(license_key="1234567890ab")  # type: ignore[call-arg]
 
         with self.assertRaises(TypeError):
-            self.client_class("47")
+            self.client_class("47")  # type: ignore
 
 
 class TestClient(TestBaseClient):
-    def setUp(self):
+    client: Client
+
+    def setUp(self) -> None:
         self.client_class = Client
         self.client = Client(42, "abcdef123456")
         self.client._base_uri = self.httpserver.url_for("/geoip/v2.1")
@@ -364,14 +408,16 @@ class TestClient(TestBaseClient):
 
 
 class TestAsyncClient(TestBaseClient):
-    def setUp(self):
+    client: AsyncClient
+
+    def setUp(self) -> None:
         self._loop = asyncio.new_event_loop()
         self.client_class = AsyncClient
         self.client = AsyncClient(42, "abcdef123456")
         self.client._base_uri = self.httpserver.url_for("/geoip/v2.1")
         self.maxDiff = 20_000
 
-    def tearDown(self):
+    def tearDown(self) -> None:
         self._loop.run_until_complete(self.client.close())
         self._loop.close()